refactor(ros-z-debug): use handle-owned cached subscriptions
